requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==4.4.2
flower==1.2.0
frozenlist==1.3.3
future==0.18.3
humanize==4.6.0
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
kombu==5.2.4
magic-filter==1.0.9
moviepy==1.0.3
multidict==6.0.4
mutagen==1.46.0
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pyaes==1.6.1
pycryptodomex==3.17
Pyrogram==2.0.104
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.29.0
six==1.16.0
TgCrypto==1.2.5
tornado==6.3.2
tqdm==4.65.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.2
yt-dlp==2023.3.4
